refactor: moved other functions out of NodeDetails.kt
